chore: move logger function to seperate package
