+++ src/cmd/root.go
@@ -4,18 +4,17 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"go.uber.org/zap"
+	"github.com/Slug-Boi/aion-cli/logger"
 )
 
 // This is an example of how to setup the logger for any CMD command you can then use it when doing calls.
 // A similar logger can be setup in any other file that requires it by importing CMD and calling the SetupLogger function
-var sugar = SetupLogger()
+var Sugar = logger.SetupLogger()
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
-	Use:   "aion", 
+	Use:   "aion",
 	Short: "A scheduling tool that takes in wishes and outputs a schedule",
-//TODO: CHANGE TO AION ACSII ART LATER
 	Long: `                                                     
  _____ _                 _ _ 
 |  _  |_|___ ___ ___ ___| |_|
@@ -41,17 +40,5 @@ func Execute() {
 }
 
 func init() {
-	
-}
 
-// The setup logger function will live in the root command as most logging should be propagated up to the CMD commands
-// This allows us to create a local logger for each command that can be used to log errors and info messages
-func SetupLogger() *zap.SugaredLogger {
-	// setup suggered zap logger
-	// https://github.com/uber-go/zap
-	logger, _ := zap.NewProduction()
-	defer logger.Sync() // flushes buffer, if any
-	sugar := logger.Sugar()
-
-	return sugar
 }
